requirements.txt
aionefit
paho-mqtt
bosch-thermostat-http-client
bosch-thermostat-client
